moved display_messages in common folder
